Move music related info to a ViewModel to persist configuration changes.
